Add marker resource to instrumented multi-release JARs
Gradle can only instrument bytecode up to the version ASM understands.
For multi-release JARs this means that versioned directories for the
unsupported JVM versions are not processed and instrumented JARs contain
no classes for these versions. However, due to the way agent-based
instrumentation works, classes are loaded first from the original JAR
and only then the agent replaces the bytecode with the instrumented
copy. Prior to this commit, it was possible that the non-instrumented
original class could be loaded if Gradle would run on the unsupported
JVM.

Now instrumentation puts a marker resource into processed multi-release
JARs which contains instrumentation status. Root directory always has
"instrumented" status. If there is an unsupported versioned directory
with class files, it gets a versioned resource with "not instrumented"
status.

The replacing class loader loads the resource and checks the status when
transformed classes are looked up in the JAR for the first time. If the
JVM is unsupported and there are versioned class for this JVM, the class
loader loads the "not instrumented" version of the resource and aborts
loading the class.

Unfortunately, we cannot use manifest attributes instead of the marker
resource, because we have to use streaming API to process JARs and the
order of entries is undefined.